+++ CHANGELOG.md
@@ -10,6 +10,7 @@
  - DeepEventedModel for eventing on all state changes
  - OpenAI Chat Completions API request/response objects
  - New Transaction system w/ Evented data objects for tracking changes
+ - Better comparison system
 
 ## [0.2.5] - 2025-05-28
  - 98% test coverage